Add access to TimeZone internals for swift-corelibs-foundation
